# apps/api/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SentinelGlobal API",
    "version": "1.0.0",
    "description": "API pública de monitoramento global de riscos em tempo real. Ideal para integração com empresas, órgãos de defesa civil e aplicativos.",
    "contact": {
      "name": "SentinelGlobal",
      "url": "https://sentinelglobal.io"
    }
  },
  "servers": [
    { "url": "https://api.sentinelglobal.io/v1", "description": "Produção" },
    { "url": "http://localhost:3001", "description": "Desenvolvimento" }
  ],
  "security": [
    { "ApiKeyAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Chave de API obtida em /api-keys"
      }
    },
...
  },
  "paths": {
...
    "/webhooks": {
      "post": {
        "summary": "Criar webhook para receber eventos",
        "tags": ["Webhooks"],
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": { "type": "string", "format": "uri" },
                  "events": { "type": "array", "items": { "type": "string" } },
                  "minRiskLevel": { "type": "integer", "default": 3 },
                  "countries": { "type": "array", "items": { "type": "string" } }
                },
                "required": ["url", "events"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Webhook criado" } }
      },
      "get": {
        "summary": "Listar webhooks",
        "tags": ["Webhooks"],
        "security": [{ "ApiKeyAuth": [] }],
        "responses": { "200": { "description": "Lista de webhooks" } }
      }
    },
...
  }
}